+++ src/cmd/6c/div.c
@@ -41,7 +41,7 @@
 #define	T32	TN(32)
 
 int
-multiplier(ulong d, int p, uvlong *mp)
+multiplier(uint32 d, int p, uvlong *mp)
 {
 	int l;
 	uvlong mlo, mhi, tlo, thi;
@@ -67,7 +67,7 @@ multiplier(ulong d, int p, uvlong *mp)
 }
 
 int
-sdiv(ulong d, ulong *mp, int *sp)
+sdiv(uint32 d, uint32 *mp, int *sp)
 {
 	int s;
 	uvlong m;
@@ -82,7 +82,7 @@ sdiv(ulong d, ulong *mp, int *sp)
 }
 
 int
-udiv(ulong d, ulong *mp, int *sp, int *pp)
+udiv(uint32 d, uint32 *mp, int *sp, int *pp)
 {
 	int p, s;
 	uvlong m;
@@ -113,7 +113,7 @@ void
 sdivgen(Node *l, Node *r, Node *ax, Node *dx)
 {
 	int a, s;
-	ulong m;
+	uint32 m;
 	vlong c;
 
 	c = r->vconst;
@@ -137,7 +137,7 @@ void
 udivgen(Node *l, Node *r, Node *ax, Node *dx)
 {
 	int a, s, t;
-	ulong m;
+	uint32 m;
 	Node nod;
 
 	a = udiv(r->vconst, &m, &s, &t);
@@ -184,7 +184,7 @@ sext(Node *d, Node *s, Node *l)
 }
 
 void
-sdiv2(long c, int v, Node *l, Node *n)
+sdiv2(int32 c, int v, Node *l, Node *n)
 {
 	Node nod;
 
@@ -206,7 +206,7 @@ sdiv2(long c, int v, Node *l, Node *n)
 }
 
 void
-smod2(long c, int v, Node *l, Node *n)
+smod2(int32 c, int v, Node *l, Node *n)
 {
 	Node nod;
